feat(runbotics-scheduler): :sparkles: Secures update and delete enpoints with PrivilageTypes

+++ runbotics/runbotics-scheduler/src/scheduler-database/credential/credential.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCredentialDto } from './dto/create-credential.dto';
 import { UpdateCredentialDto } from './dto/update-credential.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -16,6 +16,7 @@ import { CredentialNotifyMailArgs, CredentialOperationType } from './credential.
 import { Specs } from '#/utils/specification/specifiable.decorator';
 import { Paging } from '#/utils/page/pageable.decorator';
 import { getPage } from '#/utils/page/page';
+import { CredentialCollection } from '../credential-collection/credential-collection.entity';
 
 const RELATIONS = ['attributes', 'createdBy', 'collection.credentialCollectionUser'];
 
@@ -44,7 +45,7 @@ export class CredentialService {
     );
 
     if (!hasEditCollectionAccess) {
-      throw new NotFoundException('You do not have access to edit this collection');
+      throw new ForbiddenException('You do not have access to edit this collection');
     }
 
     if (!template) {
@@ -260,6 +261,11 @@ export class CredentialService {
       throw new NotFoundException(`Could not find credential with id ${id}`);
     }
 
+    const collection = await this.collectionService.findOneAccessibleById(credential.collectionId, user);
+    this.throwExceptionIfUserHasReadAcces(collection, user);
+
+    console.log('I go here');
+
     const credentialToUpdate = this.credentialRepo.create({
       ...credential,
       ...credentialDto,
@@ -297,6 +303,13 @@ export class CredentialService {
   async removeById(id: string, user: IUser) {
     const credential = await this.findById(id, user.tenantId);
 
+    if (!credential) {
+      throw new NotFoundException(`Could not find credential with id ${id}`);
+    }
+
+    const collection = await this.collectionService.findOneAccessibleById(credential.collectionId, user);
+    this.throwExceptionIfUserHasReadAcces(collection, user);
+
     return this.credentialRepo.remove(credential).then((credential) => {
         this.notifyCredentialCollectionOwner({
             executor: user,
@@ -314,7 +327,10 @@ export class CredentialService {
       throw new NotFoundException(`Could not find credential with id ${id}`);
     }
 
-        const attribute = credential.attributes.find(attr => attr.name === attributeName);
+    const collection = await this.collectionService.findOneAccessibleById(credential.collectionId, user);
+    this.throwExceptionIfUserHasReadAcces(collection, user);
+
+    const attribute = credential.attributes.find(attr => attr.name === attributeName);
 
     const encryptedValue =
       this.secretService.encrypt(attributeDto.value, user.tenantId);
@@ -437,5 +453,17 @@ export class CredentialService {
                 color: collection.color
             }
         };
+    }
+
+    private throwExceptionIfUserHasReadAcces(collection: CredentialCollection, user: IUser) {
+        const hasEditCollectionAccess = collection.credentialCollectionUser
+            .some(collectionUser =>
+              collectionUser.userId === user.id &&
+              collectionUser.privilegeType === PrivilegeType.WRITE
+        );
+
+        if (!hasEditCollectionAccess) {
+            throw new ForbiddenException('You cannot modify this collection');
+        }
     }
 }
